Updating report task to be on par with sbt plugin
In sbt-scoverage plugin, report task is much more detailed. This commit
updates gradle version of report task to be on par with it's sbt
counterpart.

Updated version of task brings 4 new configuration parameters:

 - coverageOutputCobertura
 - coverageOutputXML
 - coverageOutputHTML
 - coverageDebug

All could be used to enable/disable different outputs. All but last one
defaults to true. Readme is updated with this information.

Code in `ScoverageWriter.java` is re-written from sbt-scoverage and
converted to Java. It could be moved to scalac-scoverage-runtime to be
used by all plugins. It's not inlined in `SingleReportApp.java` because
it'll be used also in ScoverageAggregate task (in next PR).

If there is no test data in project report task now only issues a warning
and not crash (it's important for multi-project setups without test
code/code to test in some sub-projects).